改进：时间精度Int->Float

+++ app/src/main/java/com/application/traverldiary/models/Destination.kt
@@ -2,7 +2,7 @@ package com.application.traverldiary.models
 
 class Destination(
     var name: String,
-    fromTime: Int,
-    toTime: Int,
+    fromTime: Float,
+    toTime: Float,
     isHappened: Boolean
 ) : Journey(nameOrId = name, fromTime, toTime, isHappened)
